+++ test/app/components/missing_collection_parameter_name_component.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class MissingCollectionParameterNameComponent < ViewComponent::Base
+  with_collection_parameter :foo
+
+  def initialize(bar:); end
+end